+++ index.d.ts
@@ -1,1 +1,2 @@
 export { default as BarChart } from "./BarChart.svelte";
+export { default as ScatterGraph } from "./ScatterGraph.svelte";