+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # **Changelog**
 
+## **v1.1.0**
+
+-   [feat] added scatter graph
+-   [fix] fixed pill rounding on smaller bars
+
 ## **v1.0.1**
 
 -   [fix] fixed transition on page navigation